third_party/blink/web_tests/external/wpt/app-history/navigate-event/transitionWhile-on-synthetic-event.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_dom("SecurityError", () => new AppHistoryNavigateEvent("navigate").transitionWhile(Promise.resolve()));
}, "AppHistoryNavigateEvent's transitionWhile() throws if invoked on a synthetic event");
</script>
